Revised error responds in iiif.cc + synchronized with upstream
- Improved error responding for corrupted files (files with width or height 1px are reported as corrupted)

- Changed all error responds to XML format as in specification

- Added changes from upstream (improved rotation algorithm, improved 32bit support)

+++ src/Transforms.h
@@ -77,6 +77,6 @@ void filter_interpolate_bilinear( RawTile& in, unsigned int w, unsigned int h );
     @param angle angle of rotation - currently only rotations by 90, 180 and 270 degrees
     are suported, for other values, no rotation will occur
 */
-void filter_rotate( RawTile& in, double angle );
+void filter_rotate( RawTile& in, float angle );
 
 #endif
